fixed data loss issue

+++ signal_processing/salsa.h
@@ -6,7 +6,7 @@
 
 typedef struct
 {
-    float *times;
+    double *times;
     double *frameTot;
     int frameRate;
     int numFrames;
